+++ svd-rs/src/enumeratedvalues.rs
@@ -152,7 +152,21 @@ impl EnumeratedValues {
         Ok(())
     }
     /// Get the usage of these enumerated values.
-    pub fn usage(&self) -> Usage {
-        self.usage.unwrap_or_default()
+    pub fn usage(&self) -> Option<Usage> {
+        if self.derived_from.is_some() {
+            None
+        } else {
+            Some(self.usage.unwrap_or_default())
+        }
+    }
+
+    /// Get `enumeratedValue` by name
+    pub fn get_value(&self, name: &str) -> Option<&EnumeratedValue> {
+        self.values.iter().find(|e| e.name == name)
+    }
+
+    /// Get mutable `enumeratedValue` by name
+    pub fn get_mut_value(&mut self, name: &str) -> Option<&mut EnumeratedValue> {
+        self.values.iter_mut().find(|e| e.name == name)
     }
 }